+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to
 - 🔧(backend) allow configuring settings OIDC_OP_USER_ENDPOINT_FORMAT
 - ⚡️(helm) create a dedicated svc and deployment for yprovider converter #2368
 - ✨(backend) allow to leave a document #2365
+- 📈(backend) create a utils to capture event with posthog
 
 ### Changed